+++ cli/Sources/TuistConfig/TuistGeneratedProjectOptions.swift
@@ -49,7 +49,8 @@ public struct TuistGeneratedProjectOptions: Equatable, Hashable {
                 includeGenerateScheme: false,
                 enableCaching: false,
                 registryEnabled: false,
-                warningsAsErrors: .none
+                warningsAsErrors: .none,
+                defaultSwiftVersion: GenerationOptions.defaultSwiftVersionValue
             ),
             installOptions: .init(passthroughSwiftPackageManagerArguments: []),
             cacheOptions: CacheOptions(
@@ -60,6 +61,8 @@ public struct TuistGeneratedProjectOptions: Equatable, Hashable {
     }
 
     public struct GenerationOptions: Codable, Hashable, Equatable {
+        public static let defaultSwiftVersionValue = "5"
+
         public enum StaticSideEffectsWarningTargets: Codable, Hashable, Equatable {
             case all
             case none
@@ -105,6 +108,7 @@ public struct TuistGeneratedProjectOptions: Equatable, Hashable {
         public let enableCaching: Bool
         public let registryEnabled: Bool
         public let warningsAsErrors: WarningsAsErrors
+        public let defaultSwiftVersion: String
 
         public init(
             resolveDependenciesWithSystemScm: Bool,
@@ -121,7 +125,8 @@ public struct TuistGeneratedProjectOptions: Equatable, Hashable {
             includeGenerateScheme: Bool,
             enableCaching: Bool = false,
             registryEnabled: Bool = false,
-            warningsAsErrors: WarningsAsErrors = .none
+            warningsAsErrors: WarningsAsErrors = .none,
+            defaultSwiftVersion: String = GenerationOptions.defaultSwiftVersionValue
         ) {
             self.resolveDependenciesWithSystemScm = resolveDependenciesWithSystemScm
             self.disablePackageVersionLocking = disablePackageVersionLocking
@@ -138,6 +143,7 @@ public struct TuistGeneratedProjectOptions: Equatable, Hashable {
             self.enableCaching = enableCaching
             self.registryEnabled = registryEnabled
             self.warningsAsErrors = warningsAsErrors
+            self.defaultSwiftVersion = defaultSwiftVersion
         }
 
         #if DEBUG
@@ -157,7 +163,8 @@ public struct TuistGeneratedProjectOptions: Equatable, Hashable {
                 includeGenerateScheme: Bool = true,
                 enableCaching: Bool = false,
                 registryEnabled: Bool = false,
-                warningsAsErrors: TuistGeneratedProjectOptions.GenerationOptions.WarningsAsErrors = .none
+                warningsAsErrors: TuistGeneratedProjectOptions.GenerationOptions.WarningsAsErrors = .none,
+                defaultSwiftVersion: String = GenerationOptions.defaultSwiftVersionValue
             ) -> Self {
                 .init(
                     resolveDependenciesWithSystemScm: resolveDependenciesWithSystemScm,
@@ -174,7 +181,8 @@ public struct TuistGeneratedProjectOptions: Equatable, Hashable {
                     includeGenerateScheme: includeGenerateScheme,
                     enableCaching: enableCaching,
                     registryEnabled: registryEnabled,
-                    warningsAsErrors: warningsAsErrors
+                    warningsAsErrors: warningsAsErrors,
+                    defaultSwiftVersion: defaultSwiftVersion
                 )
             }
         #endif