+++ server/lib/tuist_web/live/user_reset_password_live.ex
@@ -40,18 +40,18 @@ defmodule TuistWeb.UserResetPasswordLive do
             <.form data-part="form" for={@form} id="reset_password_form" phx-submit="reset_password">
               <.text_input
                 field={@form[:password]}
+                type="password"
                 input_type="password"
                 label={dgettext("dashboard_auth", "New password")}
                 show_prefix={false}
-                show_suffix={false}
                 required
               />
               <.text_input
                 field={@form[:password_confirmation]}
+                type="password"
                 input_type="password"
                 label={dgettext("dashboard_auth", "Confirm password")}
                 show_prefix={false}
-                show_suffix={false}
                 required
               />
               <.button